refactor(step3): product package로 이동
