+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -2,6 +2,7 @@ package kitchenpos.domain;
 
 import javax.persistence.*;
 import java.util.UUID;
+import kitchenpos.product.domain.Product;
 
 @Table(name = "menu_product")
 @Entity
