+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,13 +1,18 @@
-package kitchenpos.ui;
-
-import kitchenpos.application.ProductService;
-import kitchenpos.domain.Product;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+package kitchenpos.product.ui;
 
 import java.net.URI;
 import java.util.List;
 import java.util.UUID;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.domain.Product;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RequestMapping("/api/products")
 @RestController
